ADD - Gerenciamento de produtos

+++ Controllers/CategoriaController.cs
@@ -11,25 +11,15 @@ namespace LojaNemesis.Controllers
 {
   [Authorize]
   [Route("api/[controller]")]
-  public class CategoriaController : Controller
+  public class CategoriaController : BaseController
   {
-    private AppDbContext context;
-    public CategoriaController(AppDbContext context)
-    {
-      this.context = context;
-    }
+    public CategoriaController(AppDbContext context) : base(context) { }
 
     [HttpGet]
-    public List<Categoria> Get()
-    {
-      return context.Categoria.ToList();
-    }
+    public IActionResult Get() => Ok(context.Categoria.ToList());
 
     [HttpGet("{id}")]
-    public Categoria Get(int id)
-    {
-      return context.Categoria.FirstOrDefault(p => p.Id == id);
-    }
+    public IActionResult Get(int id) => Ok(context.Categoria.FirstOrDefault(p => p.Id == id));
 
     [HttpPost]
     public void Post([FromBody] Categoria categoria)
@@ -45,6 +35,8 @@ namespace LojaNemesis.Controllers
     [HttpDelete("{id}")]
     public void Delete(int id)
     {
+      if (context.Produto.FirstOrDefault(p => p.Categoria.Id == id) != null)
+        throw new ValidateException("Existem produtos utilizando esta categoria.");
       var c = context.Categoria.FirstOrDefault(p => p.Id == id);
       context.Categoria.Remove(c);
       context.SaveChanges();
